backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
python-multipart
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
email-validator
razorpay
